Refactoring: - 'webviews' folder renamed to 'views'
