+++ src/views/IntegrationsView.ts
