+++ src/views/NodeConfigView.ts
